Started creating a page/module for users with supplier role.
