+++ inventory_system/procurement_view/admin.py
@@ -1,12 +1,9 @@
 from django.contrib import admin
-from .models import RequestQuotation, RequestQuotationItem, QuotationSubmission, QuotationSubmissionItem, PurchaseOrder, PurchaseOrderItem
+from .models import RequestQuotation, RequestQuotationItem, PurchaseOrder, PurchaseOrderItem
 
 # Register your models here.
 admin.site.register(RequestQuotation)
 admin.site.register(RequestQuotationItem)
-
-admin.site.register(QuotationSubmission)
-admin.site.register(QuotationSubmissionItem)
 
 admin.site.register(PurchaseOrder)
 admin.site.register(PurchaseOrderItem)
