+++ inventory_system/procurement_view/forms.py
@@ -2,7 +2,7 @@ import os
 from django import forms
 from django.forms import modelformset_factory
 from django.contrib.auth.models import User
-from .models import RequestQuotation, RequestQuotationItem, QuotationSubmission, QuotationSubmissionItem, PurchaseOrder, PurchaseOrderItem
+from .models import RequestQuotation, RequestQuotationItem, PurchaseOrder, PurchaseOrderItem
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -14,7 +14,7 @@ class RequestQuotationForm(forms.ModelForm):
     employee = forms.ModelChoiceField(queryset=User.objects.filter(userpermission__role=USER_ROLE), label=USER_ROLE)
 
     class Meta:
-        STATUS_CHOICES = [('Ongoing', 'Ongoing'), ('Ended', 'Ended')]
+        STATUS_CHOICES = [(status, status) for status in os.getenv('RQ_STATUS_CHOICES', '').split(',')]
         
         model = RequestQuotation
         
@@ -34,34 +34,6 @@ class RequestQuotationItemForm(forms.ModelForm):
 RequestQuotationItemFormSet = modelformset_factory(RequestQuotationItem, form=RequestQuotationItemForm, extra=5)
 
 
-class QuotationSubmissionForm(forms.ModelForm):
-    USER_ROLE = os.environ.get('ROLE_3')
-
-    supplier = forms.ModelChoiceField(queryset=User.objects.filter(userpermission__role=USER_ROLE), label=USER_ROLE)
-    
-    class Meta:
-        STATUS_CHOICES = [('Accept', 'Accept'), ('Pending', 'Pending')]
-
-        model = QuotationSubmission
-        fields = [
-            'supplier', 'buyer_company_name', 'buyer_address', 'buyer_contact', 
-            'quotation_no', 'prepared_by', 'quote_valid_until',  
-            'terms_and_conditions', 'total_amount', 'status'
-        ]
-        
-        widgets = {
-            'quote_valid_until': forms.DateInput(attrs={'type': 'date'}),
-            'date_submitted': forms.DateInput(attrs={'type': 'date'}),
-            'status': forms.Select(choices=STATUS_CHOICES),
-        }
-
-class QuotationSubmissionItemForm(forms.ModelForm):
-    class Meta:
-        model = QuotationSubmissionItem
-        fields = ['product_name', 'quantity', 'unit_price']
-
-QuotationSubmissionItemFormSet = modelformset_factory(QuotationSubmissionItem, form=QuotationSubmissionItemForm, extra=5)
-
 
 class PurchaseOrderForm(forms.ModelForm):
     USER_ROLE = os.environ.get('ROLE_3')
@@ -69,7 +41,7 @@ class PurchaseOrderForm(forms.ModelForm):
     supplier = forms.ModelChoiceField(queryset=User.objects.filter(userpermission__role=USER_ROLE), label=USER_ROLE)
 
     class Meta:
-        STATUS_CHOICES = [('Pending', 'Pending'), ('Approved', 'Approved'), ('Rejected', 'Rejected')]
+        STATUS_CHOICES = [(status, status) for status in os.getenv('PO_STATUS_CHOICES', '').split(',')]
         
         model = PurchaseOrder
         fields = [
