+++ inventory_system/procurement_view/models.py
@@ -32,36 +32,6 @@ class RequestQuotationItem(models.Model):
     def __str__(self):
         return f"{self.product_name} - {self.quantity} units"
 
-class QuotationSubmission(models.Model):
-    supplier = models.ForeignKey(User, on_delete=models.CASCADE)
-    buyer_company_name = models.CharField(max_length=255)
-    buyer_address = models.TextField()
-    buyer_contact = models.CharField(max_length=255)
-    quotation_no = models.CharField(max_length=50)
-    prepared_by = models.CharField(max_length=255)
-    quote_valid_until = models.DateField()
-    date_submitted = models.DateField(auto_now_add=True)
-    terms_and_conditions = models.TextField()
-    total_amount = models.DecimalField(max_digits=10, decimal_places=2)
-    status = models.CharField(max_length=50)
-
-    def __str__(self):
-        return f"{self.quotation_no} - {self.buyer_company_name}"
-    
-    def save(self, *args, **kwargs):
-        if not self.quotation_no:
-            self.quotation_no = generate_unique_procurement_no("QS", QuotationSubmission)
-
-        super(QuotationSubmission, self).save(*args, **kwargs)
-
-class QuotationSubmissionItem(models.Model):
-    quotation_submission = models.ForeignKey(QuotationSubmission, on_delete=models.CASCADE, related_name='items')
-    product_name = models.CharField(max_length=255)
-    quantity = models.PositiveIntegerField()
-    unit_price = models.DecimalField(max_digits=10, decimal_places=2)
-
-    def __str__(self):
-        return f"{self.product_name} (Qty: {self.quantity})"
 
 class PurchaseOrder(models.Model):
     supplier = models.ForeignKey(User, on_delete=models.CASCADE, blank=True, null=True)
@@ -77,12 +47,6 @@ class PurchaseOrder(models.Model):
 
     def __str__(self):
         return f"PO #{self.quotation_no} - {self.supplier}"
-    
-    def save(self, *args, **kwargs):
-        if not self.quotation_no:
-            self.quotation_no = generate_unique_procurement_no("PO", PurchaseOrder)
-
-        super(PurchaseOrder, self).save(*args, **kwargs)
 
 class PurchaseOrderItem(models.Model):
     purchase_order = models.ForeignKey(PurchaseOrder, on_delete=models.CASCADE)
@@ -91,4 +55,4 @@ class PurchaseOrderItem(models.Model):
     unit_price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
-        return f"{self.product} (PO #{self.purchase_order.po_no})"
+        return f"{self.product} (PO #{self.purchase_order.quotation_no})"
