+++ inventory_system/supplier_view/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import QuotationSubmission, QuotationSubmissionItem
+
+
+admin.site.register(QuotationSubmission)
+admin.site.register(QuotationSubmissionItem)
