+++ inventory_system/supplier_view/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class SupplierViewConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'supplier_view'
