+++ inventory_system/supplier_view/forms.py
@@ -1,0 +1,37 @@
+import os
+from django import forms
+from django.forms import modelformset_factory
+from django.contrib.auth.models import User
+from .models import QuotationSubmission, QuotationSubmissionItem
+from dotenv import load_dotenv
+
+load_dotenv()
+
+
+class QuotationSubmissionForm(forms.ModelForm):
+    USER_ROLE = os.environ.get('ROLE_3')
+
+    supplier = forms.ModelChoiceField(queryset=User.objects.filter(userpermission__role=USER_ROLE), label=USER_ROLE)
+    
+    class Meta:
+        STATUS_CHOICES = [('Accept', 'Accept'), ('Pending', 'Pending')]
+
+        model = QuotationSubmission
+        fields = [
+            'supplier', 'buyer_company_name', 'buyer_address', 'buyer_contact', 
+            'quotation_no', 'prepared_by', 'quote_valid_until',  
+            'terms_and_conditions', 'total_amount', 'status'
+        ]
+        
+        widgets = {
+            'quote_valid_until': forms.DateInput(attrs={'type': 'date'}),
+            'date_submitted': forms.DateInput(attrs={'type': 'date'}),
+            'status': forms.Select(choices=STATUS_CHOICES),
+        }
+
+class QuotationSubmissionItemForm(forms.ModelForm):
+    class Meta:
+        model = QuotationSubmissionItem
+        fields = ['product_name', 'quantity', 'unit_price']
+
+QuotationSubmissionItemFormSet = modelformset_factory(QuotationSubmissionItem, form=QuotationSubmissionItemForm, extra=5)
