+++ inventory_system/supplier_view/templates/create_quotation_submission.html
@@ -1,4 +1,4 @@
-{% extends 'procurement_base.html' %}
+{% extends 'supplier_base.html' %}
 
 {% block title %}Create Quotation Submission{% endblock %}
