+++ inventory_system/supplier_view/templates/supplier_base.html
@@ -2,17 +2,26 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>{% block title %} {% endblock %}</title>
+    <title>{% block title %}{% endblock %}</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
 </head>
 <body>
-    
-    <div class="container">
-        {% block content %}
-        {% endblock %}
+    {% include 'navigation_bar.html' %}
+
+    <div class="container-fluid">
+        <div class="row">
+            {% include 'supplier_side_navigation.html' %}
+
+            <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4" style="height: 93vh; overflow: auto; scroll-width: thin;">
+                {% block content %}
+                {% endblock %}
+            </main>
+            
+        </div>
     </div>
 
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.4/dist/umd/popper.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
 </body>
 </html>
