+++ inventory_system/supplier_view/templates/supplier_side_navigation.html
@@ -1,0 +1,17 @@
+<nav class="col-md-2 d-md-block bg-primary text-white sidebar" style="position: sticky; top: 0; height: 93vh;">
+    <div class="position-sticky">
+        <ul class="nav flex-column">
+
+            <li class="nav-item">
+                <a class="nav-link text-white" href="{% url 'request_quotations_list' %}">
+                    <strong>View Request Quotations</strong>
+                </a>
+
+                <a class="nav-link text-white" href="{% url 'create_quotation_submission' %}">
+                    <strong>View Purchase Orders</strong>
+                </a>
+            </li>
+            
+        </ul>
+    </div>
+</nav>
